requirements.txt
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==7.0.0
cryptography==42.0.7
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.1
